Remove things from broken java config
Having those enabled in the jdtls config resulting in permission issues
when the editor or the jdtls program tried to read the files/libs there.

An easy work-around is to install the JDKs somewhere else, using a
different tool than the supplied archlinux-java program to manage JDKs.
Or a different tool to manage JDKs in a location that's not bound to the
root user and group.